projects/claude_generated/python/p_5/p_5_requirements_claude.txt
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
numpy==1.24.3
